Ansible/Artifactory - Improve offline upgrades
Improve play to better handle nodes without internet access. By pinning
tasks to occur when the tar is downloaded or unarchived it allows for
more coverage of scenarios.

Similarly, check to see if the jdbc driver provided as part of the tar
is the version set in the variables.

This allows for users to provide the tar in presteps to calling the role
if they choose / need to.

Zachary.Wayer@garmin.com